Reorganize all the text functions
